requirements.txt
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-xml==1.4.0
drf-tracking==1.5.0
gunicorn==20.0.4
psycopg2==2.8.5
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
